+++ src/DbInterface.cpp
@@ -23,8 +23,6 @@
 
 #include <algorithm>
 #include <tuple>
-#include <vector>
-#include <functional>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/bind/bind.hpp>
@@ -325,9 +323,6 @@ void DbInterface::initialize()
         mAppDbForwardingCommandTablePtr = std::make_shared<swss::Table> (
             mAppDbPtr.get(), APP_FORWARDING_STATE_COMMAND_TABLE_NAME
         );
-        mAppDbIcmpEchoSessionTablePtr = std::make_shared<swss::ProducerStateTable> (
-            mAppDbPtr.get(), APP_ICMP_ECHO_SESSION_TABLE_NAME
-        );
         mStateDbMuxLinkmgrTablePtr = std::make_shared<swss::Table> (
             mStateDbPtr.get(), STATE_MUX_LINKMGR_TABLE_NAME
         );
@@ -339,9 +334,6 @@ void DbInterface::initialize()
         );
         mStateDbSwitchCauseTablePtr = std::make_shared<swss::Table> (
             mStateDbPtr.get(),  STATE_MUX_SWITCH_CAUSE_TABLE_NAME
-        );
-        mStateDbIcmpEchoSessionTablePtr = std::make_shared<swss::Table> (
-            mStateDbPtr.get(),  STATE_ICMP_ECHO_SESSION_TABLE_NAME
         );
         mMuxStateTablePtr = std::make_shared<swss::Table> (mStateDbPtr.get(), STATE_MUX_CABLE_TABLE_NAME);
 
@@ -814,47 +806,6 @@ void DbInterface::getPortCableType(std::shared_ptr<swss::DBConnector> configDbCo
 
     configDbMuxCableTable.getContent(entries);
     processPortCableType(entries);
-}
-
-//
-// ---> processLinkFailureDetectionType(std::vector<swss::KeyOpFieldsValuesTuple> &entries);
-//
-// process Mux Cable Table enteries to get linkFailureDetectionType by defaut its software
-//
-void DbInterface::processLinkFailureDetectionType(std::vector<swss::KeyOpFieldsValuesTuple> &entries)
-{
-    for (auto &entry: entries) {
-        std::string portName = kfvKey(entry);
-        std::string operation = kfvOp(entry);
-        std::vector<swss::FieldValueTuple> fieldValues = kfvFieldsValues(entry);
-
-        std::string field = "link_failure_detection_type";
-        std::vector<swss::FieldValueTuple>::const_iterator cit = std::find_if(
-            fieldValues.cbegin(),
-            fieldValues.cend(),
-            [&field] (const swss::FieldValueTuple &fv) {return fvField(fv) == field;}
-        );
-        std::string linkFailureDetectionType = (cit != fieldValues.cend() ? cit->second : "software");
-
-        MUXLOGDEBUG(boost::format("port: %s, %s = %s") % portName % field % linkFailureDetectionType);
-
-        mMuxManagerPtr->updateLinkFailureDetectionType(portName, linkFailureDetectionType);
-    }
-}
-
-//
-// ---> getPortCableType(std::shared_ptr<swss::DBConnector> configDbConnector);
-//
-// retrieve the Link Failure Detection Type (HW/SW) from config
-//
-void DbInterface::getLinkFailureDetectionType(std::shared_ptr<swss::DBConnector> configDbConnector)
-{
-    MUXLOGINFO("Reading link_failure_detection_type");
-    swss::Table configDbMuxCableTable(configDbConnector.get(), CFG_MUX_CABLE_TABLE_NAME);
-    std::vector<swss::KeyOpFieldsValuesTuple> entries;
-
-    configDbMuxCableTable.getContent(entries);
-    processLinkFailureDetectionType(entries);
 }
 
 //
@@ -1574,128 +1525,6 @@ void DbInterface::processTsaEnableNotification(std::deque<swss::KeyOpFieldsValue
 }
 
 //
-// ---> createIcmpEchoSession(std::string key, IcmpHwOffloadEntriesPtr entries);
-//
-// triggers creation of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-//
-void DbInterface::createIcmpEchoSession(std::string key, IcmpHwOffloadEntriesPtr entries)
-{
-    MUXLOGDEBUG(key);
-
-    boost::asio::post(mStrand, boost::bind(
-        &DbInterface::handleIcmpEchoSession,
-        this,
-        key,
-        entries.release()
-    ));
-}
-
-//
-// ---> handleIcmpEchoSession(std::string key, IcmpHwOffloadEntries *entries);
-//
-// handles creation of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-//
-void DbInterface::handleIcmpEchoSession(std::string key, IcmpHwOffloadEntries *entries)
-{
-    std::vector<swss::FieldValueTuple> fvs;
-    for(auto& entry : *entries) {
-        const std::string field = entry.first;
-        const std::string value = entry.second;
-        fvs.emplace_back(field, value);
-        MUXLOGDEBUG(boost::format("APP_ICMP_ECHO_SESSION_TABLE::key: %s, field: %s, value: %s") %
-                key %
-                field %
-                value
-            );
-    }
-    mAppDbIcmpEchoSessionTablePtr->set(key, fvs);
-    delete entries;
-}
-
-//
-// ---> deleteIcmpEchoSession(std::string key);
-//
-// handles deletion of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-//
-void DbInterface::deleteIcmpEchoSession(std::string key)
-{
-    MUXLOGDEBUG(boost::format("APP_ICMP_ECHO_SESSION_TABLE::key: %s ") %
-            key
-        );
-    mAppDbIcmpEchoSessionTablePtr->del(key);
-}
-
-//
-// ---> extractIfnameAndSessionType(const std::string &key, std::string &ifname, std::string &sessionType);
-//
-// helper function to extract interface name and session_type from the key
-//
-void DbInterface::extractIfnameAndSessionType(const std::string &key, std::string &ifname, std::string &sessionType)
-{
-    size_t firstColon = key.find(':');
-    size_t secondColon = key.find(':', firstColon + 1);
-    size_t thirdColon = key.find(':', secondColon + 1);
-    size_t fourthColon = key.find(':', thirdColon + 1);
-
-    if (firstColon != std::string::npos &&
-        secondColon != std::string::npos &&
-        thirdColon != std::string::npos &&
-        fourthColon != std::string::npos) {
-        ifname = key.substr(secondColon + 1, thirdColon - secondColon - 1);
-        sessionType = key.substr(fourthColon + 1);
-    } else {
-        std::cerr << "Unexpected format: " << key << std::endl;
-    }
-}
-
-//
-// ---> processIcmpEchoSessionStateNotification(std::deque<swss::KeyOpFieldsValuesTuple> &entries);
-//
-// process each entery of ICMP_ECHO_SESSION_TABLE extract Port, Session_Type(Normal/Rx) and state itself
-//
-void DbInterface::processIcmpEchoSessionStateNotification(std::deque<swss::KeyOpFieldsValuesTuple> &entries)
-{
-      for (auto &entry: entries) {
-        std::string key = kfvKey(entry);
-        std::string port;
-        std::string session_type;
-        extractIfnameAndSessionType(key, port, session_type);
-        std::string opertation = kfvOp(entry);
-        std::vector<swss::FieldValueTuple> fieldValues = kfvFieldsValues(entry);
-
-        std::vector<swss::FieldValueTuple>::const_iterator cit = std::find_if(
-            fieldValues.cbegin(),
-            fieldValues.cend(),
-            [] (const swss::FieldValueTuple &fv) {return fvField(fv) == "state";}
-        );
-        if (cit != fieldValues.cend()) {
-            const std::string field = cit->first;
-            const std::string value = cit->second;
-
-            MUXLOGDEBUG(boost::format("port: %s, f: %s, v: %s") %
-                port %
-                field %
-                value
-            );
-            mMuxManagerPtr->updateLinkFailureDetectionState(port, value, session_type);
-        }
-    }
-}
-
-//
-// ---> handleIcmpEchoSessionStateNotification(swss::SubscriberStateTable &stateDbIcmpEchoSessionTable);
-//
-// extract enteries from State_DB ICMP_ECHO_SESSION_TABLE and call processing function
-//
-void DbInterface::handleIcmpEchoSessionStateNotification(swss::SubscriberStateTable &stateDbIcmpEchoSessionTable)
-{
-    std::deque<swss::KeyOpFieldsValuesTuple> entries;
-
-    stateDbIcmpEchoSessionTable.pops(entries);
-    processIcmpEchoSessionStateNotification(entries);
-}
-
-//
 // ---> handleSwssNotification();
 //
 // main thread method for handling SWSS notification
@@ -1727,14 +1556,11 @@ void DbInterface::handleSwssNotification()
     swss::SubscriberStateTable stateDbMuxInfoTable(stateDbPtr.get(), MUX_CABLE_INFO_TABLE);
     // for getting peer's admin forwarding state
     swss::SubscriberStateTable stateDbPeerMuxTable(stateDbPtr.get(), STATE_PEER_HW_FORWARDING_STATE_TABLE_NAME);
-    // for getting icmp echo session state
-    swss::SubscriberStateTable stateDbIcmpEchoSessionTable(stateDbPtr.get(), STATE_ICMP_ECHO_SESSION_TABLE_NAME);
 
     getTorMacAddress(configDbPtr);
     getVlanNames(configDbPtr);
     getLoopback2InterfaceInfo(configDbPtr);
     getPortCableType(configDbPtr);
-    getLinkFailureDetectionType(configDbPtr);
     getServerIpAddress(configDbPtr);
     getSoCIpAddress(configDbPtr);
 
@@ -1758,7 +1584,6 @@ void DbInterface::handleSwssNotification()
     swssSelect.addSelectable(&stateDbMuxInfoTable);
     swssSelect.addSelectable(&stateDbPeerMuxTable);
     swssSelect.addSelectable(&netlinkNeighbor);
-    swssSelect.addSelectable(&stateDbIcmpEchoSessionTable);
 
     while (mPollSwssNotifcation) {
         swss::Selectable *selectable;
@@ -1794,8 +1619,6 @@ void DbInterface::handleSwssNotification()
             handlePeerLinkStateNotification(stateDbMuxInfoTable);
         } else if (selectable == static_cast<swss::Selectable *> (&stateDbPeerMuxTable)) {
             handlePeerMuxStateNotification(stateDbPeerMuxTable);
-        } else if (selectable == static_cast<swss::Selectable *> (&stateDbIcmpEchoSessionTable)) {
-            handleIcmpEchoSessionStateNotification(stateDbIcmpEchoSessionTable);
         } else if (selectable == static_cast<swss::Selectable *> (&netlinkNeighbor)) {
             continue;
         } else {
