+++ src/DbInterface.h
@@ -27,7 +27,6 @@
 #include <map>
 #include <memory>
 
-#include <utility>
 #include <boost/thread.hpp>
 #include <boost/thread/barrier.hpp>
 #include <boost/date_time/posix_time/posix_time.hpp>
@@ -36,10 +35,9 @@
 #include "swss/producerstatetable.h"
 #include "swss/subscriberstatetable.h"
 #include "swss/warm_restart.h"
-#include "link_prober/LinkProberBase.h"
+
 #include "link_manager/LinkManagerStateMachineActiveStandby.h"
 #include "mux_state/MuxState.h"
-
 
 namespace test {
 class MuxManagerTest;
@@ -56,15 +54,10 @@ namespace mux
 #define APP_PEER_HW_FORWARDING_STATE_TABLE_NAME    "HW_FORWARDING_STATE_PEER"
 #define STATE_PEER_HW_FORWARDING_STATE_TABLE_NAME   "HW_MUX_CABLE_TABLE_PEER"
 
-#define STATE_ICMP_ECHO_SESSION_TABLE_NAME "ICMP_ECHO_SESSION_TABLE"
-#define APP_ICMP_ECHO_SESSION_TABLE_NAME "ICMP_ECHO_SESSION_TABLE"
-
 #define STATE_MUX_SWITCH_CAUSE_TABLE_NAME "MUX_SWITCH_CAUSE"
 
 class MuxManager;
 using ServerIpPortMap = std::map<boost::asio::ip::address, std::string>;
-using IcmpHwOffloadEntries = std::vector<std::pair<std::string, std::string>>;
-using IcmpHwOffloadEntriesPtr = std::unique_ptr<IcmpHwOffloadEntries>;
 
 /**
  *@class DbInterface
@@ -403,24 +396,6 @@ public:
     *@return port to mux mode map
     */
     virtual std::map<std::string, std::string> getMuxModeConfig();
-
-    /**
-    *@method createIcmpEchoSession
-    *
-    *@brief retrieve mux mode config
-    *
-    *@return triggers creation of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-    */
-    void createIcmpEchoSession(std::string key, IcmpHwOffloadEntriesPtr entries);
-
-    /**
-    *@method deleteIcmpEchoSession
-    *
-    *@brief retrieve mux mode config
-    *
-    *@return handles deletion of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-    */
-    void deleteIcmpEchoSession(std::string key);
 
 private:
     friend class test::MuxManagerTest;
@@ -883,66 +858,12 @@ private:
 
     /**
      * @method processTsaEnableNotification
-     *
+     * 
      * @brief process Tsa Enable Notification
-     *
+     * 
      * @return none
      */
     void processTsaEnableNotification(std::deque<swss::KeyOpFieldsValuesTuple> &entries);
-
-    /**
-     * @method handleIcmpEchoSessionStateNotification
-     *
-     * @brief extract enteries from State_DB ICMP_ECHO_SESSION_TABLE and call processing function
-     *
-     * @return none
-     */
-    void handleIcmpEchoSessionStateNotification(swss::SubscriberStateTable &stateDbIcmpEchoSessionTable);
-
-    /**
-     * @method handleIcmpEchoSession
-     *
-     * @brief handles creation of a ICMP_ECHO_SESSION in APP_ICMP_ECHO_SESSION_TABLE
-     *
-     * @return none
-     */
-    void handleIcmpEchoSession(std::string key, IcmpHwOffloadEntries *entries);
-
-    /**
-     * @method processIcmpEchoSessionStateNotification
-     *
-     * @brief process each entery of ICMP_ECHO_SESSION_TABLE extract Port, Session_Type(Normal/Rx) and state itself
-     *
-     * @return none
-     */
-    void processIcmpEchoSessionStateNotification(std::deque<swss::KeyOpFieldsValuesTuple> &entries);
-
-    /**
-     * @method processLinkFailureDetectionType
-     *
-     * @brief process Mux Cable Table enteries to get linkFailureDetectionType by defaut its software
-     *
-     * @return none
-     */
-    void processLinkFailureDetectionType(std::vector<swss::KeyOpFieldsValuesTuple> &entries);
-
-    /**
-     * @method getLinkFailureDetectionType
-     *
-     * @brief retrieve the Link Failure Detection Type (HW/SW) from config
-     *
-     * @return none
-     */
-    void getLinkFailureDetectionType(std::shared_ptr<swss::DBConnector> configDbConnector);
-
-    /**
-     * @method extractIfnameAndSessionType
-     *
-     * @brief helper function to extract interface name and session_type from the key
-     *
-     * @return none
-     */
-    void extractIfnameAndSessionType(const std::string &key, std::string &ifname, std::string &sessionType);
 
 private:
     static std::vector<std::string> mMuxState;
@@ -959,10 +880,8 @@ private:
     std::shared_ptr<swss::DBConnector> mStateDbPtr;
     std::shared_ptr<swss::Table> mMuxStateTablePtr;
 
-    // for communicating with orchagent for mux stae
+    // for communicating with orchagent
     std::shared_ptr<swss::ProducerStateTable> mAppDbMuxTablePtr;
-    // for communicating with orchagent for icmp echo session
-    std::shared_ptr<swss::ProducerStateTable> mAppDbIcmpEchoSessionTablePtr;
     // for communication with driver (setting peer's forwarding state)
     std::shared_ptr<swss::Table> mAppDbPeerMuxTablePtr;
     // for communicating with the driver (probing the mux)
@@ -977,8 +896,6 @@ private:
     std::shared_ptr<swss::Table> mStateDbLinkProbeStatsTablePtr;
     // for writing mux switch reason to state db 
     std::shared_ptr<swss::Table> mStateDbSwitchCauseTablePtr;
-    // for reading icmp echo session state from state db
-    std::shared_ptr<swss::Table> mStateDbIcmpEchoSessionTablePtr;
 
     std::shared_ptr<boost::thread> mSwssThreadPtr;
