+++ src/link_manager/LinkManagerStateMachineActiveStandby.cpp
@@ -23,7 +23,6 @@
 
 #include <boost/bind/bind.hpp>
 
-#include "link_prober/LinkProberSw.h"
 #include "link_manager/LinkManagerStateMachineActiveStandby.h"
 #include "common/MuxLogger.h"
 #include "common/MuxException.h"
@@ -348,49 +347,49 @@ void ActiveStandbyStateMachine::handleSwssBladeIpv4AddressUpdate(boost::asio::ip
         mMuxPortConfig.setBladeIpv4Address(address);
 
         try {
-            mLinkProberPtr = std::make_shared<link_prober::LinkProberSw> (
+            mLinkProberPtr = std::make_shared<link_prober::LinkProber> (
                 mMuxPortConfig,
                 getStrand().context(),
                 mLinkProberStateMachinePtr.get()
             );
             mInitializeProberFnPtr = boost::bind(
-                &link_prober::LinkProberBase::initialize, mLinkProberPtr.get()
+                &link_prober::LinkProber::initialize, mLinkProberPtr.get()
             );
             mStartProbingFnPtr = boost::bind(
-                &link_prober::LinkProberBase::startProbing, mLinkProberPtr.get()
+                &link_prober::LinkProber::startProbing, mLinkProberPtr.get()
             );
             mUpdateEthernetFrameFnPtr = boost::bind(
-                &link_prober::LinkProberBase::updateEthernetFrame, mLinkProberPtr.get()
+                &link_prober::LinkProber::updateEthernetFrame, mLinkProberPtr.get()
             );
             mProbePeerTorFnPtr = boost::bind(
-                &link_prober::LinkProberBase::probePeerTor, mLinkProberPtr.get()
+                &link_prober::LinkProber::probePeerTor, mLinkProberPtr.get()
             );
             mDetectLinkFnPtr = boost::bind(
-                &link_prober::LinkProberBase::detectLink, mLinkProberPtr.get()
+                &link_prober::LinkProber::detectLink, mLinkProberPtr.get()
             );
             mSuspendTxFnPtr = boost::bind(
-                &link_prober::LinkProberBase::suspendTxProbes, mLinkProberPtr.get(), boost::placeholders::_1
+                &link_prober::LinkProber::suspendTxProbes, mLinkProberPtr.get(), boost::placeholders::_1
             );
             mResumeTxFnPtr = boost::bind(
-                &link_prober::LinkProberBase::resumeTxProbes, mLinkProberPtr.get()
+                &link_prober::LinkProber::resumeTxProbes, mLinkProberPtr.get()
             );
             mSendPeerSwitchCommandFnPtr = boost::bind(
-                &link_prober::LinkProberBase::sendPeerSwitchCommand, mLinkProberPtr.get()
+                &link_prober::LinkProber::sendPeerSwitchCommand, mLinkProberPtr.get()
             );
             mResetIcmpPacketCountsFnPtr = boost::bind(
-                &link_prober::LinkProberBase::resetIcmpPacketCounts, mLinkProberPtr.get()
+                &link_prober::LinkProber::resetIcmpPacketCounts, mLinkProberPtr.get()
             );
             mShutdownTxFnPtr = boost::bind(
-                &link_prober::LinkProberBase::shutdownTxProbes, mLinkProberPtr.get()
+                &link_prober::LinkProber::shutdownTxProbes, mLinkProberPtr.get()
             );
             mRestartTxFnPtr = boost::bind(
-                &link_prober::LinkProberBase::restartTxProbes, mLinkProberPtr.get()
+                &link_prober::LinkProber::restartTxProbes, mLinkProberPtr.get()
             );
             mDecreaseIntervalFnPtr = boost::bind(
-                &link_prober::LinkProberBase::decreaseProbeIntervalAfterSwitch, mLinkProberPtr.get(), boost::placeholders::_1
+                &link_prober::LinkProber::decreaseProbeIntervalAfterSwitch, mLinkProberPtr.get(), boost::placeholders::_1
             );
             mRevertIntervalFnPtr = boost::bind(
-                &link_prober::LinkProberBase::revertProbeIntervalAfterSwitchComplete, mLinkProberPtr.get()
+                &link_prober::LinkProber::revertProbeIntervalAfterSwitchComplete, mLinkProberPtr.get()
             );
             mComponentInitState.set(LinkProberComponent);
