+++ src/link_prober/IcmpPayload.cpp
@@ -34,6 +34,7 @@ namespace link_prober
 //
 // static members
 //
+boost::uuids::uuid IcmpPayload::mGuid;
 uint32_t IcmpPayload::mCookie = 0x47656d69;
 uint32_t IcmpPayload::mVersion = 0;
 
@@ -47,7 +48,20 @@ IcmpPayload::IcmpPayload() :
     version(htonl(mVersion)),
     seq(0)
 {
+    memcpy(uuid, mGuid.data, sizeof(uuid));
+}
 
+//
+// ---> generateGuid()
+//
+// generate GUID for the current instance of linkmgrd
+//
+void IcmpPayload::generateGuid()
+{
+    boost::uuids::random_generator gen;
+    mGuid = gen();
+
+    MUXLOGWARNING(boost::format("Link Prober generated GUID: {%s}") % boost::uuids::to_string(mGuid));
 }
 
 } /* namespace link_prober */
