+++ src/link_prober/IcmpPayload.h
@@ -31,7 +31,7 @@
 #include <netinet/ip.h>
 #include <netinet/ip_icmp.h>
 #include <net/ethernet.h>
-#include <unordered_set>
+
 #include <boost/uuid/uuid.hpp>
 
 #include <sys/types.h>
@@ -39,12 +39,6 @@
 #define MUX_MAX_ICMP_BUFFER_SIZE    9100
 
 __BEGIN_DECLS
-
-#include <endian.h>
-inline uint64_t ntohll(uint64_t x)
-{
-    return be64toh(x);
-}
 
 namespace link_prober
 {
@@ -104,6 +98,33 @@ struct IcmpPayload {
     IcmpPayload();
 
     /**
+    *@method generateGuid
+    *
+    *@brief generate GUID for the current instance of linkmgrd
+    *
+    *@return none
+    */
+    static void generateGuid();
+
+    /**
+    *@method getGuidData
+    *
+    *@brief getter for GUID data
+    *
+    *@return pointer to current GUID data
+    */
+    static uint8_t* getGuidData() {return reinterpret_cast<uint8_t *> (mGuid.data);};
+
+    /**
+    *@method getGuid
+    *
+    *@brief getter for GUID object
+    *
+    *@return reference to current GUID object
+    */
+    static boost::uuids::uuid& getGuid() {return mGuid;};
+
+    /**
     *@method getCookie
     *
     *@brief getter for ICMP cookie
@@ -124,7 +145,7 @@ struct IcmpPayload {
 private:
     static uint32_t mCookie;
     static uint32_t mVersion;
-
+    static boost::uuids::uuid mGuid;
 } __attribute__((packed));
 
 static_assert(sizeof(IcmpPayload) % 2 == 0,
