+++ src/link_prober/LinkProberState.cpp
@@ -63,12 +63,6 @@ LinkProberState* LinkProberState::handleEvent(IcmpUnknownEvent &event)
     return nullptr;
 }
 
-LinkProberState* LinkProberState::handleEvent(IcmpWaitEvent &event)
-{
-    MUXLOGERROR(getMuxPortConfig().getPortName());
-    return nullptr;
-}
-
 LinkProberState* LinkProberState::handleEvent(IcmpPeerActiveEvent &event)
 {
     MUXLOGERROR(getMuxPortConfig().getPortName());
@@ -76,12 +70,6 @@ LinkProberState* LinkProberState::handleEvent(IcmpPeerActiveEvent &event)
 }
 
 LinkProberState* LinkProberState::handleEvent(IcmpPeerUnknownEvent &event)
-{
-    MUXLOGERROR(getMuxPortConfig().getPortName());
-    return nullptr;
-}
-
-LinkProberState* LinkProberState::handleEvent(IcmpPeerWaitEvent &event)
 {
     MUXLOGERROR(getMuxPortConfig().getPortName());
     return nullptr;
