+++ src/link_prober/LinkProberStateMachineActiveActive.cpp
@@ -28,7 +28,6 @@ namespace link_prober
 //          link_manager::LinkManagerStateMachineBase &linkManagerStateMachinePtr,
 //          boost::asio::io_service::strand &strand,
 //          common::MuxPortConfig &muxPortConfig,
-//          bool proberTypeHardware,
 //          LinkProberState::Label label
 //      );
 //
@@ -188,7 +187,7 @@ void LinkProberStateMachineActiveActive::processEvent(SuspendTimerExpiredEvent &
 void LinkProberStateMachineActiveActive::processEvent(MuxProbeRequestEvent &muxProbeRequestEvent)
 {
     boost::asio::io_service::strand &strand = mLinkManagerStateMachinePtr->getStrand();
-    boost::asio::post(strand, boost::bind(
+    boost::asio::post(mStrand, boost::bind(
         &link_manager::LinkManagerStateMachineBase::handleMuxProbeRequestEvent,
         mLinkManagerStateMachinePtr
     ));
