+++ src/link_prober/LinkProberStateMachineActiveStandby.h
@@ -21,8 +21,8 @@
  *      Author: tamer
  */
 
-#ifndef LINK_PROBER_LINKPROBERSTATEMACHINEACTIVESTANDBY_H_
-#define LINK_PROBER_LINKPROBERSTATEMACHINEACTIVESTANDBY_H_
+#ifndef LINK_PROBER_LINKPROBERSTATEMACHINE_H_
+#define LINK_PROBER_LINKPROBERSTATEMACHINE_H_
 
 #include "link_prober/LinkProberStateMachineBase.h"
